requirements.txt
fastapi
uvicorn
pillow
zplgrf
python-multipart
jinja2